Wave Spawner, Enemy Deletion, Ship Rotation
The wave spawner is now in the game but there are a couple of bugs. The wave spawner will only spawn the first wave. It also complains that it can't find the enemy tag when there are no enemies available.

Enemies are now destroyed if they travel off screen.

The ship is now locked to the middle of the screen and rotates to look at the mouse pointer.